Changed the distance results back to float - too much pain with org.sygmetrics library

+++ src/main/java/uk/ac/standrews/cs/utilities/m_tree/Distance.java
@@ -18,5 +18,5 @@ package uk.ac.standrews.cs.utilities.m_tree;
 
 public interface Distance<T> {
 
-    double distance(T a, T b);
+    float distance(T a, T b);
 }
